refactor: move some parts of ConfigScreen to a new screens.config package
