+++ core/src/com/agateau/pixelwheels/screens/ConfigScreen.java
@@ -27,12 +27,9 @@ import com.agateau.pixelwheels.LogExporter;
 import com.agateau.pixelwheels.PwGame;
 import com.agateau.pixelwheels.PwRefreshHelper;
 import com.agateau.pixelwheels.VersionInfo;
-import com.agateau.pixelwheels.gameinput.GameInputHandlerFactories;
 import com.agateau.pixelwheels.gameinput.GameInputHandlerFactory;
-import com.agateau.pixelwheels.gameinput.GamepadInputHandler;
-import com.agateau.pixelwheels.gameinput.KeyboardInputHandler;
+import com.agateau.pixelwheels.screens.config.InputSelectorController;
 import com.agateau.pixelwheels.utils.StringUtils;
-import com.agateau.ui.UiAssets;
 import com.agateau.ui.anchor.AnchorGroup;
 import com.agateau.ui.menu.ButtonMenuItem;
 import com.agateau.ui.menu.LabelMenuItem;
@@ -43,18 +40,13 @@ import com.agateau.ui.menu.SelectorMenuItem;
 import com.agateau.ui.menu.SwitchMenuItem;
 import com.agateau.ui.menu.TabMenuItem;
 import com.agateau.ui.uibuilder.UiBuilder;
-import com.agateau.utils.Assert;
 import com.agateau.utils.FileUtils;
 import com.agateau.utils.PlatformUtils;
 import com.agateau.utils.log.NLog;
-import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
-import com.badlogic.gdx.scenes.scene2d.utils.Drawable;
-import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
-import com.badlogic.gdx.utils.Array;
 
 /** The config screen */
 public class ConfigScreen extends PwStageScreen {
@@ -71,10 +63,6 @@ public class ConfigScreen extends PwStageScreen {
     MenuItemGroup mLanguageGroup;
 
     private boolean mLanguageChanged = false;
-
-    interface GameInputHandlerConfigScreenFactory {
-        Screen createScreen(PwGame game, int playerIdx);
-    }
 
     public ConfigScreen(PwGame game) {
         super(game.getAssets().ui);
@@ -303,75 +291,6 @@ public class ConfigScreen extends PwStageScreen {
         return screen;
     }
 
-    class InputSelectorController {
-        private final SelectorMenuItem<GameInputHandlerFactory> mSelector;
-        private final ButtonMenuItem mConfigureButton;
-        private final int mPlayerIdx;
-
-        InputSelectorController(
-                SelectorMenuItem<GameInputHandlerFactory> selector,
-                ButtonMenuItem configureButton,
-                int idx) {
-            mSelector = selector;
-            mConfigureButton = configureButton;
-            mPlayerIdx = idx;
-
-            UiAssets uiAssets = mGame.getAssets().ui;
-            Array<GameInputHandlerFactory> inputFactories =
-                    GameInputHandlerFactories.getAvailableFactories();
-            for (GameInputHandlerFactory factory : inputFactories) {
-                String iconName = "input-icons/" + factory.getId();
-                Drawable drawable = new TextureRegionDrawable(uiAssets.atlas.findRegion(iconName));
-                selector.addEntry(drawable, factory.getName(), factory);
-            }
-
-            selector.getActor()
-                    .addListener(
-                            new ChangeListener() {
-                                @Override
-                                public void changed(ChangeEvent event, Actor actor) {
-                                    onInputChanged();
-                                }
-                            });
-
-            configureButton.addListener(
-                    new MenuItemListener() {
-                        @Override
-                        public void triggered() {
-                            onConfigureClicked();
-                        }
-                    });
-        }
-
-        private void onInputChanged() {
-            GameInputHandlerFactory factory = mSelector.getCurrentData();
-            mGame.getConfig().setPlayerInputHandlerFactory(mPlayerIdx, factory);
-            mGame.getConfig().flush();
-            updateConfigureButton();
-        }
-
-        private void onConfigureClicked() {
-            GameInputHandlerFactory factory = mSelector.getCurrentData();
-            GameInputHandlerConfigScreenFactory configScreenFactory =
-                    getInputConfigScreenFactory(factory);
-            Assert.check(configScreenFactory != null, "No config screen for this game factory");
-            mGame.pushScreen(configScreenFactory.createScreen(mGame, mPlayerIdx));
-        }
-
-        private void setStartupState() {
-            GameInputHandlerFactory factory =
-                    mGame.getConfig().getPlayerInputHandlerFactory(mPlayerIdx);
-            mSelector.setCurrentData(factory);
-            updateConfigureButton();
-        }
-
-        private void updateConfigureButton() {
-            GameInputHandlerFactory factory = mSelector.getCurrentData();
-            boolean canBeConfigured = getInputConfigScreenFactory(factory) != null;
-            mConfigureButton.setDisabled(!canBeConfigured);
-        }
-    }
-
     private void setupInputSelector(Menu menu, MenuItemGroup group, String label, final int idx) {
         SelectorMenuItem<GameInputHandlerFactory> selector = new SelectorMenuItem<>(menu);
         group.addItemWithLabel(label, selector);
@@ -380,19 +299,8 @@ public class ConfigScreen extends PwStageScreen {
         group.addItemWithLabel("", configureButton);
 
         InputSelectorController controller =
-                new InputSelectorController(selector, configureButton, idx);
+                new InputSelectorController(mGame, selector, configureButton, idx);
         controller.setStartupState();
-    }
-
-    private GameInputHandlerConfigScreenFactory getInputConfigScreenFactory(
-            GameInputHandlerFactory factory) {
-        if (factory instanceof GamepadInputHandler.Factory) {
-            return GamepadConfigScreen::new;
-        } else if (factory instanceof KeyboardInputHandler.Factory) {
-            return KeyboardConfigScreen::new;
-        } else {
-            return null;
-        }
     }
 
     @Override
