+++ core/src/com/agateau/pixelwheels/screens/config/GameInputHandlerConfigScreenFactory.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2023 Aurélien Gâteau <mail@agateau.com>
+ *
+ * This file is part of Pixel Wheels.
+ *
+ * Pixel Wheels is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License as published by the Free
+ * Software Foundation, either version 3 of the License, or (at your option)
+ * any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along with
+ * this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.agateau.pixelwheels.screens.config;
+
+import com.agateau.pixelwheels.PwGame;
+import com.badlogic.gdx.Screen;
+
+/** A factory to create a config screen for an input method */
+public interface GameInputHandlerConfigScreenFactory {
+    Screen createScreen(PwGame game, int playerIdx);
+}
